requirements.api.txt
# Core API Dependencies
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
redis==5.0.1
SQLAlchemy==1.4.49

# ML & Data Science
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
lightgbm==4.6.0
joblib==1.3.2
torch==2.1.0
transformers==4.46.3

# Visualization (FIXED: Added missing seaborn)
matplotlib==3.8.2
seaborn==0.13.0

# MLflow
mlflow==2.9.2

# NLP
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
textstat==0.7.10

# Utilities
requests==2.32.5
tqdm==4.66.1
python-json-logger==2.0.7
praw==7.7.1

# Monitoring
psutil==7.1.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.28.1